Optimization fixes
Set garage door to not open on workdays before 3AM.
Turn geysers OFF when stove is ON.
Created Hue light group with WInHue, and removed the light group created in HA (and updated google and homekit entities).
Added sensors to show current and latest HASSIO versions and sensor to notify on UI if running latest version.
Set sun boolean to change to trigger every 2 minutes instead of when sun rises or sets.
Set light inside garage to turn OFF when it is day, regardless of whether the door is open or closed (previously only when closed).
Three new automations for when a geyser starts using power, it turns the other geysers OFF (different automations will turn it back ON when it stops using power).